requirements.txt
fastapi
uvicorn
redis
python-dotenv
sqlmodel
psycopg2-binary
alembic